ipacm: Add LAN Stats support for static policy clients
Update IPA_HANDLE_WAN_VLAN_PDN_UP to check for
lan stats enabled when static policy mode is enabled.
If lan stats is enabled, then IPACM will modify the
modem UL(2nd pass) rules to have stats enabled.

+++ ipacm/inc/IPACM_Wan.h
@@ -125,7 +125,7 @@ typedef struct
 	uint32_t ipv4_addr;
 	bool wan_up_vlan;
 	bool is_xlat;
-	uint8_t associated_VIDs[IPA_MAX_NUM_SW_PDNS];
+	uint16_t associated_VIDs[IPA_MAX_NUM_SW_PDNS];
 	uint8_t VID_cnt = 0;
 	IPACM_Wan *pIface;
 }ipacm_ipv4_wan_iface;
@@ -134,7 +134,7 @@ typedef struct
 {
 	uint32_t ipv6_prefix[2];
 	bool wan_up_vlan_v6;
-	uint8_t associated_VIDs[IPA_MAX_NUM_SW_PDNS];
+	uint16_t associated_VIDs[IPA_MAX_NUM_SW_PDNS];
 	uint8_t VID_cnt = 0;
 	IPACM_Wan *pIface;
 }ipacm_ipv6_wan_iface;
